feat: Make event creation endpoint infer EventID from event CAR file data (#341)
