+++ api/Cargo.toml
@@ -28,6 +28,7 @@ swagger.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 prometheus-client.workspace = true
+hex.workspace = true
 
 [target.'cfg(target_os = "linux")'.dependencies]
 jemalloc_pprof = "0.1.0"
