+++ api/src/server/event.rs
@@ -1,9 +1,8 @@
 use std::collections::HashMap;
 
 use anyhow::{anyhow, bail, Context, Result};
-use ceramic_core::{EventId, Network};
+use ceramic_core::{Cid, EventId, Network};
 use ceramic_event::unvalidated;
-use cid::Cid;
 use ipld_core::ipld::Ipld;
 use iroh_car::CarReader;
 use tokio::io::AsyncRead;
@@ -123,104 +122,14 @@ mod tests {
 
     use super::*;
 
+    use crate::tests::{
+        decode_multibase_str, mock_get_init_event, MockAccessModelStoreTest, DATA_EVENT_CAR,
+        INIT_EVENT_CAR, TIME_EVENT_CAR,
+    };
     use async_trait::async_trait;
     use expect_test::{expect, Expect};
     use mockall::{mock, predicate};
     use tracing_test::traced_test;
-
-    const INIT_EVENT_CID: &str = "bagcqcerar2aga7747dm6fota3iipogz4q55gkaamcx2weebs6emvtvie2oha";
-    const INIT_EVENT_PAYLOAD_CID: &str =
-        "bafyreiaroclcgqih242byss6pneufencrulmeex2ttfdzefst67agwq3im";
-    const INIT_EVENT_CAR: &str = "
-        uO6Jlcm9vdHOB2CpYJgABhQESII6AYH_8-NniumDaEPcbPId6ZQAMFfViEDLxGVnVBNOOZ3ZlcnNpb24
-        B0QEBcRIgEXCWI0EH1zQcSl57SUKRoo0WwhL6nMo8kLKfvgNaG0OiZGRhdGGhZXN0ZXBoGQFNZmhlYWR
-        lcqRjc2VwZW1vZGVsZW1vZGVsWCjOAQIBhQESIKDoMqM144vTQLQ6DwKZvzxRWg_DPeTNeRCkPouTHo1
-        YZnVuaXF1ZUxEpvE6skELu2qFaN5rY29udHJvbGxlcnOBeDhkaWQ6a2V5Ono2TWt0QnluQVBMckV5ZVM
-        3cFZ0aGJpeVNjbWZ1OG41Vjdib1hneHlvNXEzU1pSUroCAYUBEiCOgGB__PjZ4rpg2hD3GzyHemUADBX
-        1YhAy8RlZ1QTTjqJncGF5bG9hZFgkAXESIBFwliNBB9c0HEpee0lCkaKNFsIS-pzKPJCyn74DWhtDanN
-        pZ25hdHVyZXOBomlwcm90ZWN0ZWRYgXsiYWxnIjoiRWREU0EiLCJraWQiOiJkaWQ6a2V5Ono2TWt0Qnl
-        uQVBMckV5ZVM3cFZ0aGJpeVNjbWZ1OG41Vjdib1hneHlvNXEzU1pSUiN6Nk1rdEJ5bkFQTHJFeWVTN3B
-        WdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlIifWlzaWduYXR1cmVYQCQDjlx8fT8rbTR4088HtOE
-        27LJMc38DSuf1_XtK14hDp1Q6vhHqnuiobqp5EqNOp0vNFCCzwgG-Dsjmes9jJww";
-    const INIT_EVENT: &str = "
-        uomdwYXlsb2FkWCQBcRIgEXCWI0EH1zQcSl57SUKRoo0WwhL6nMo8kLKfvgNaG0Nqc2lnbmF0dXJlc4G
-        iaXByb3RlY3RlZFiBeyJhbGciOiJFZERTQSIsImtpZCI6ImRpZDprZXk6ejZNa3RCeW5BUExyRXllUzd
-        wVnRoYml5U2NtZnU4bjVWN2JvWGd4eW81cTNTWlJSI3o2TWt0QnluQVBMckV5ZVM3cFZ0aGJpeVNjbWZ
-        1OG41Vjdib1hneHlvNXEzU1pSUiJ9aXNpZ25hdHVyZVhAJAOOXHx9PyttNHjTzwe04TbsskxzfwNK5_X
-        9e0rXiEOnVDq-Eeqe6KhuqnkSo06nS80UILPCAb4OyOZ6z2MnDA";
-    const INIT_EVENT_PAYLOAD: &str = "
-        uomRkYXRhoWVzdGVwaBkBTWZoZWFkZXKkY3NlcGVtb2RlbGVtb2RlbFgozgECAYUBEiCg6DKjNeOL00C
-        0Og8Cmb88UVoPwz3kzXkQpD6Lkx6NWGZ1bmlxdWVMRKbxOrJBC7tqhWjea2NvbnRyb2xsZXJzgXg4ZGl
-        kOmtleTp6Nk1rdEJ5bkFQTHJFeWVTN3BWdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlI";
-
-    const DATA_EVENT_CAR: &str = "
-        uO6Jlcm9vdHOB2CpYJgABhQESICddBxl5Sk2e7I20pzX9kDLf0jj6WvIQ1KqbM3WQiClDZ3ZlcnNpb24
-        BqAEBcRIgdtssXEgR7sXQQQA1doBpxUpTn4pcAaVFZfQjyo-03SGjYmlk2CpYJgABhQESII6AYH_8-Nn
-        iumDaEPcbPId6ZQAMFfViEDLxGVnVBNOOZGRhdGGBo2JvcGdyZXBsYWNlZHBhdGhmL3N0ZXBoZXZhbHV
-        lGQFOZHByZXbYKlgmAAGFARIgjoBgf_z42eK6YNoQ9xs8h3plAAwV9WIQMvEZWdUE0466AgGFARIgJ10
-        HGXlKTZ7sjbSnNf2QMt_SOPpa8hDUqpszdZCIKUOiZ3BheWxvYWRYJAFxEiB22yxcSBHuxdBBADV2gGn
-        FSlOfilwBpUVl9CPKj7TdIWpzaWduYXR1cmVzgaJpcHJvdGVjdGVkWIF7ImFsZyI6IkVkRFNBIiwia2l
-        kIjoiZGlkOmtleTp6Nk1rdEJ5bkFQTHJFeWVTN3BWdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlI
-        jejZNa3RCeW5BUExyRXllUzdwVnRoYml5U2NtZnU4bjVWN2JvWGd4eW81cTNTWlJSIn1pc2lnbmF0dXJ
-        lWECym-Kwb5ti-T5dCygt4zf8Lr6MescAbkk_DILoy3fFjYG8fZVUCGKDQiTTHbNbzOk1yze7-2hA3AK
-        dBfzJY1kA";
-
-    const TIME_EVENT_CAR: &str = "
-        uOqJlcm9vdHOB2CpYJQABcRIgcmqgb7eHSgQ32hS1NGVKZruLJGcKDI1f4lqOyNYn3eVndmVyc2lvbgG
-        3AQFxEiByaqBvt4dKBDfaFLU0ZUpmu4skZwoMjV_iWo7I1ifd5aRiaWTYKlgmAAGFARIgjoBgf_z42eK
-        6YNoQ9xs8h3plAAwV9WIQMvEZWdUE045kcGF0aGEwZHByZXbYKlgmAAGFARIgJ10HGXlKTZ7sjbSnNf2
-        QMt_SOPpa8hDUqpszdZCIKUNlcHJvb2bYKlglAAFxEiAFKLx3fi7-yD1aPNyqnblI_r_5XllReVz55jB
-        MvMxs9q4BAXESIAUovHd-Lv7IPVo83KqduUj-v_leWVF5XPnmMEy8zGz2pGRyb2902CpYJQABcRIgfWt
-        bF-FQN6GN6ZL8OtHvp2YrGlmLbZwkOl6UY-3AUNFmdHhIYXNo2CpYJgABkwEbIBv-WU6fLnsyo5_lDST
-        C_T-xUlW95brOAUDByGHJzbCRZnR4VHlwZWpmKGJ5dGVzMzIpZ2NoYWluSWRvZWlwMTU1OjExMTU1MTE
-        xeQFxEiB9a1sX4VA3oY3pkvw60e-nZisaWYttnCQ6XpRj7cBQ0YPYKlgmAAGFARIgJ10HGXlKTZ7sjbS
-        nNf2QMt_SOPpa8hDUqpszdZCIKUP22CpYJQABcRIgqVOMo-IVjo08Mk0cim3Z8flNyHY7c9g7uGMqeS0
-        PFHA";
-
-    mock! {
-        pub AccessModelStoreTest {}
-        #[async_trait]
-        impl AccessModelStore for AccessModelStoreTest {
-            async fn insert_many(&self, items: &[(EventId, Option<Vec<u8>>)]) -> Result<(Vec<bool>, usize)>;
-            async fn range_with_values(
-                &self,
-                start: &EventId,
-                end: &EventId,
-                offset: usize,
-                limit: usize,
-            ) -> Result<Vec<(Cid, Vec<u8>)>>;
-            async fn value_for_order_key(&self, key: &EventId) -> Result<Option<Vec<u8>>>;
-            async fn value_for_cid(&self, key: &Cid) -> Result<Option<Vec<u8>>>;
-            async fn events_since_highwater_mark(
-                &self,
-                highwater: i64,
-                limit: i64,
-            ) -> Result<(i64, Vec<Cid>)>;
-            async fn get_block(& self, cid: &Cid) -> Result<Option<Vec<u8>>>;
-        }
-    }
-
-    fn get_init_event_mock() -> MockAccessModelStoreTest {
-        // Expect two get_block calls
-        let mut mock_store = MockAccessModelStoreTest::new();
-
-        // Call to get the init event envelope
-        mock_store
-            .expect_get_block()
-            .once()
-            .with(predicate::eq(Cid::from_str(INIT_EVENT_CID).unwrap()))
-            .return_once(move |_| Ok(Some(decode_multibase_str(INIT_EVENT))));
-
-        // Call to get the init event payload
-        mock_store
-            .expect_get_block()
-            .once()
-            .with(predicate::eq(
-                Cid::from_str(INIT_EVENT_PAYLOAD_CID).unwrap(),
-            ))
-            .return_once(move |_| Ok(Some(decode_multibase_str(INIT_EVENT_PAYLOAD))));
-        mock_store
-    }
 
     async fn test_event_id_from_car(
         event_data: &str,
@@ -233,19 +142,10 @@ mod tests {
             &mock_store,
         )
         .await;
+
         expected_event_id.assert_debug_eq(&event_id);
     }
 
-    fn decode_multibase_str(encoded: &str) -> Vec<u8> {
-        let (_, bytes) = multibase::decode(
-            encoded
-                .chars()
-                .filter(|c| !c.is_whitespace())
-                .collect::<String>(),
-        )
-        .unwrap();
-        bytes
-    }
     #[tokio::test]
     #[traced_test]
     async fn event_id_from_car_signed_init_event() {
@@ -340,7 +240,9 @@ mod tests {
                 },
             )
         "#]];
-        test_event_id_from_car(DATA_EVENT_CAR, expected, get_init_event_mock()).await
+        let mut mock_model = MockAccessModelStoreTest::new();
+        mock_get_init_event(&mut mock_model);
+        test_event_id_from_car(DATA_EVENT_CAR, expected, mock_model).await
     }
     #[tokio::test]
     #[traced_test]
@@ -367,7 +269,9 @@ mod tests {
                 },
             )
         "#]];
-        test_event_id_from_car(TIME_EVENT_CAR, expected, get_init_event_mock()).await
+        let mut mock_model = MockAccessModelStoreTest::new();
+        mock_get_init_event(&mut mock_model);
+        test_event_id_from_car(TIME_EVENT_CAR, expected, mock_model).await
     }
 
     #[tokio::test]
