+++ api/src/tests.rs
@@ -22,6 +22,72 @@ use tracing_test::traced_test;
 
 struct Context;
 
+pub const INIT_EVENT_CID: &str = "bagcqcerar2aga7747dm6fota3iipogz4q55gkaamcx2weebs6emvtvie2oha";
+pub const INIT_EVENT_PAYLOAD_CID: &str =
+    "bafyreiaroclcgqih242byss6pneufencrulmeex2ttfdzefst67agwq3im";
+pub const INIT_EVENT_CAR: &str = "
+        uO6Jlcm9vdHOB2CpYJgABhQESII6AYH_8-NniumDaEPcbPId6ZQAMFfViEDLxGVnVBNOOZ3ZlcnNpb24
+        B0QEBcRIgEXCWI0EH1zQcSl57SUKRoo0WwhL6nMo8kLKfvgNaG0OiZGRhdGGhZXN0ZXBoGQFNZmhlYWR
+        lcqRjc2VwZW1vZGVsZW1vZGVsWCjOAQIBhQESIKDoMqM144vTQLQ6DwKZvzxRWg_DPeTNeRCkPouTHo1
+        YZnVuaXF1ZUxEpvE6skELu2qFaN5rY29udHJvbGxlcnOBeDhkaWQ6a2V5Ono2TWt0QnluQVBMckV5ZVM
+        3cFZ0aGJpeVNjbWZ1OG41Vjdib1hneHlvNXEzU1pSUroCAYUBEiCOgGB__PjZ4rpg2hD3GzyHemUADBX
+        1YhAy8RlZ1QTTjqJncGF5bG9hZFgkAXESIBFwliNBB9c0HEpee0lCkaKNFsIS-pzKPJCyn74DWhtDanN
+        pZ25hdHVyZXOBomlwcm90ZWN0ZWRYgXsiYWxnIjoiRWREU0EiLCJraWQiOiJkaWQ6a2V5Ono2TWt0Qnl
+        uQVBMckV5ZVM3cFZ0aGJpeVNjbWZ1OG41Vjdib1hneHlvNXEzU1pSUiN6Nk1rdEJ5bkFQTHJFeWVTN3B
+        WdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlIifWlzaWduYXR1cmVYQCQDjlx8fT8rbTR4088HtOE
+        27LJMc38DSuf1_XtK14hDp1Q6vhHqnuiobqp5EqNOp0vNFCCzwgG-Dsjmes9jJww";
+
+pub const INIT_EVENT: &str = "
+        uomdwYXlsb2FkWCQBcRIgEXCWI0EH1zQcSl57SUKRoo0WwhL6nMo8kLKfvgNaG0Nqc2lnbmF0dXJlc4G
+        iaXByb3RlY3RlZFiBeyJhbGciOiJFZERTQSIsImtpZCI6ImRpZDprZXk6ejZNa3RCeW5BUExyRXllUzd
+        wVnRoYml5U2NtZnU4bjVWN2JvWGd4eW81cTNTWlJSI3o2TWt0QnluQVBMckV5ZVM3cFZ0aGJpeVNjbWZ
+        1OG41Vjdib1hneHlvNXEzU1pSUiJ9aXNpZ25hdHVyZVhAJAOOXHx9PyttNHjTzwe04TbsskxzfwNK5_X
+        9e0rXiEOnVDq-Eeqe6KhuqnkSo06nS80UILPCAb4OyOZ6z2MnDA";
+pub const INIT_EVENT_PAYLOAD: &str = "
+        uomRkYXRhoWVzdGVwaBkBTWZoZWFkZXKkY3NlcGVtb2RlbGVtb2RlbFgozgECAYUBEiCg6DKjNeOL00C
+        0Og8Cmb88UVoPwz3kzXkQpD6Lkx6NWGZ1bmlxdWVMRKbxOrJBC7tqhWjea2NvbnRyb2xsZXJzgXg4ZGl
+        kOmtleTp6Nk1rdEJ5bkFQTHJFeWVTN3BWdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlI";
+
+pub const DATA_EVENT_CAR: &str = "
+        uO6Jlcm9vdHOB2CpYJgABhQESICddBxl5Sk2e7I20pzX9kDLf0jj6WvIQ1KqbM3WQiClDZ3ZlcnNpb24
+        BqAEBcRIgdtssXEgR7sXQQQA1doBpxUpTn4pcAaVFZfQjyo-03SGjYmlk2CpYJgABhQESII6AYH_8-Nn
+        iumDaEPcbPId6ZQAMFfViEDLxGVnVBNOOZGRhdGGBo2JvcGdyZXBsYWNlZHBhdGhmL3N0ZXBoZXZhbHV
+        lGQFOZHByZXbYKlgmAAGFARIgjoBgf_z42eK6YNoQ9xs8h3plAAwV9WIQMvEZWdUE0466AgGFARIgJ10
+        HGXlKTZ7sjbSnNf2QMt_SOPpa8hDUqpszdZCIKUOiZ3BheWxvYWRYJAFxEiB22yxcSBHuxdBBADV2gGn
+        FSlOfilwBpUVl9CPKj7TdIWpzaWduYXR1cmVzgaJpcHJvdGVjdGVkWIF7ImFsZyI6IkVkRFNBIiwia2l
+        kIjoiZGlkOmtleTp6Nk1rdEJ5bkFQTHJFeWVTN3BWdGhiaXlTY21mdThuNVY3Ym9YZ3h5bzVxM1NaUlI
+        jejZNa3RCeW5BUExyRXllUzdwVnRoYml5U2NtZnU4bjVWN2JvWGd4eW81cTNTWlJSIn1pc2lnbmF0dXJ
+        lWECym-Kwb5ti-T5dCygt4zf8Lr6MescAbkk_DILoy3fFjYG8fZVUCGKDQiTTHbNbzOk1yze7-2hA3AK
+        dBfzJY1kA";
+
+// Assumes Mainnet network
+pub const DATA_EVENT_ID: &str =
+    "ce010500aa5773c7d75777e1deb6cb4af0e69eebd504d38e0185011220275d0719794a4d9eec8db4a735fd9032dfd238fa5af210d4aa9b337590882943";
+
+pub const TIME_EVENT_CAR: &str = "
+        uOqJlcm9vdHOB2CpYJQABcRIgcmqgb7eHSgQ32hS1NGVKZruLJGcKDI1f4lqOyNYn3eVndmVyc2lvbgG
+        3AQFxEiByaqBvt4dKBDfaFLU0ZUpmu4skZwoMjV_iWo7I1ifd5aRiaWTYKlgmAAGFARIgjoBgf_z42eK
+        6YNoQ9xs8h3plAAwV9WIQMvEZWdUE045kcGF0aGEwZHByZXbYKlgmAAGFARIgJ10HGXlKTZ7sjbSnNf2
+        QMt_SOPpa8hDUqpszdZCIKUNlcHJvb2bYKlglAAFxEiAFKLx3fi7-yD1aPNyqnblI_r_5XllReVz55jB
+        MvMxs9q4BAXESIAUovHd-Lv7IPVo83KqduUj-v_leWVF5XPnmMEy8zGz2pGRyb2902CpYJQABcRIgfWt
+        bF-FQN6GN6ZL8OtHvp2YrGlmLbZwkOl6UY-3AUNFmdHhIYXNo2CpYJgABkwEbIBv-WU6fLnsyo5_lDST
+        C_T-xUlW95brOAUDByGHJzbCRZnR4VHlwZWpmKGJ5dGVzMzIpZ2NoYWluSWRvZWlwMTU1OjExMTU1MTE
+        xeQFxEiB9a1sX4VA3oY3pkvw60e-nZisaWYttnCQ6XpRj7cBQ0YPYKlgmAAGFARIgJ10HGXlKTZ7sjbS
+        nNf2QMt_SOPpa8hDUqpszdZCIKUP22CpYJQABcRIgqVOMo-IVjo08Mk0cim3Z8flNyHY7c9g7uGMqeS0
+        PFHA";
+
+/// multibase-decodes a string after stripping all whitespace characters.
+pub fn decode_multibase_str(encoded: &str) -> Vec<u8> {
+    let (_, bytes) = multibase::decode(
+        encoded
+            .chars()
+            .filter(|c| !c.is_whitespace())
+            .collect::<String>(),
+    )
+    .unwrap();
+    bytes
+}
+
 mock! {
     pub ReconInterestTest {
         fn insert(&self, key: Interest, value: Option<Vec<u8>>) -> Result<bool>;
@@ -53,79 +119,70 @@ impl AccessInterestStore for MockReconInterestTest {
         Ok(res.into_iter().map(|(k, _)| k).collect())
     }
 }
+
 mock! {
-    pub ReconModelTest {
-        fn insert_many(&self, items: &[(EventId, Option<Vec<u8>>)]) -> Result<(Vec<bool>, usize)>;
-        fn range_with_values(
+    pub AccessModelStoreTest {}
+    #[async_trait]
+    impl AccessModelStore for AccessModelStoreTest {
+        async fn insert_many(&self, items: &[(EventId, Option<Vec<u8>>)]) -> Result<(Vec<bool>, usize)>;
+        async fn range_with_values(
             &self,
             start: &EventId,
             end: &EventId,
             offset: usize,
             limit: usize,
-        ) -> Result<Vec<(Cid,Vec<u8>)>>;
-        fn value_for_order_key(&self, key: &EventId) -> Result<Option<Vec<u8>>>;
-        fn value_for_cid(&self, key: &Cid) -> Result<Option<Vec<u8>>>;
-    }
-    impl Clone for ReconModelTest {
-        fn clone(&self) -> Self;
+        ) -> Result<Vec<(Cid, Vec<u8>)>>;
+        async fn value_for_order_key(&self, key: &EventId) -> Result<Option<Vec<u8>>>;
+        async fn value_for_cid(&self, key: &Cid) -> Result<Option<Vec<u8>>>;
+        async fn events_since_highwater_mark(
+            &self,
+            highwater: i64,
+            limit: i64,
+        ) -> Result<(i64, Vec<Cid>)>;
+        async fn get_block(& self, cid: &Cid) -> Result<Option<Vec<u8>>>;
     }
 }
-#[async_trait]
-impl AccessModelStore for MockReconModelTest {
-    async fn insert_many(
-        &self,
-        items: &[(EventId, Option<Vec<u8>>)],
-    ) -> Result<(Vec<bool>, usize)> {
-        self.insert_many(items)
-    }
-    async fn range_with_values(
-        &self,
-        start: &EventId,
-        end: &EventId,
-        offset: usize,
-        limit: usize,
-    ) -> Result<Vec<(Cid, Vec<u8>)>> {
-        self.range_with_values(start, end, offset, limit)
-    }
-    async fn value_for_order_key(&self, key: &EventId) -> Result<Option<Vec<u8>>> {
-        self.value_for_order_key(key)
-    }
-    async fn value_for_cid(&self, key: &Cid) -> Result<Option<Vec<u8>>> {
-        self.value_for_cid(key)
-    }
-    async fn events_since_highwater_mark(
-        &self,
-        _highwater: i64,
-        _limit: i64,
-    ) -> anyhow::Result<(i64, Vec<Cid>)> {
-        Ok((0, vec![]))
-    }
 
-    async fn get_block(&self, _cid: &Cid) -> Result<Option<Vec<u8>>> {
-        Ok(None)
-    }
+/// Given a mock of the AccessModelStore, prepare it to expect calls to load the init event.
+pub fn mock_get_init_event(mock_store: &mut MockAccessModelStoreTest) {
+    // Expect two get_block calls
+
+    // Call to get the init event envelope
+    mock_store
+        .expect_get_block()
+        .once()
+        .with(predicate::eq(Cid::from_str(INIT_EVENT_CID).unwrap()))
+        .return_once(move |_| Ok(Some(decode_multibase_str(INIT_EVENT))));
+
+    // Call to get the init event payload
+    mock_store
+        .expect_get_block()
+        .once()
+        .with(predicate::eq(
+            Cid::from_str(INIT_EVENT_PAYLOAD_CID).unwrap(),
+        ))
+        .return_once(move |_| Ok(Some(decode_multibase_str(INIT_EVENT_PAYLOAD))));
 }
+
 #[tokio::test]
 async fn create_event() {
     let peer_id = PeerId::random();
-    let network = Network::InMemory;
-    let event_id = EventId::new(
-        &network,
-        "model",
-        &decode_multibase_data("k2t6wz4ylx0qr6v7dvbczbxqy7pqjb0879qx930c1e27gacg3r8sllonqt4xx9")
-            .unwrap(), // cspell:disable-line
-        "did:key:zGs1Det7LHNeu7DXT4nvoYrPfj3n6g7d6bj2K4AMXEvg1", // cspell:disable-line
-        &Cid::from_str("baejbeihyr3kf77etqdccjfoc33dmko2ijyugn6qk6yucfkioasjssz3bbu").unwrap(), // cspell:disable-line
-        &Cid::from_str("baejbeicqtpe5si4qvbffs2s7vtbk5ccbsfg6owmpidfj3zeluqz4hlnz6m").unwrap(), // cspell:disable-line
-    );
-    let event_id_str = multibase::encode(Base::Base16Lower, event_id.to_bytes());
-    let event_data = "f".to_string();
+    let network = Network::Mainnet;
+    let expected_event_id = EventId::try_from(hex::decode(DATA_EVENT_ID).unwrap()).unwrap();
+
+    // Remove whitespace from event CAR file
+    let event_data = DATA_EVENT_CAR
+        .chars()
+        .filter(|c| !c.is_whitespace())
+        .collect::<String>();
     let mock_interest = MockReconInterestTest::new();
-    let mut mock_model = MockReconModelTest::new();
+    let mut mock_model = MockAccessModelStoreTest::new();
+    mock_get_init_event(&mut mock_model);
     let args = vec![(
-        event_id.clone(),
-        Some(decode_multibase_data(event_data.as_str()).unwrap()),
+        expected_event_id,
+        Some(decode_multibase_data(&event_data).unwrap()),
     )];
+
     mock_model
         .expect_insert_many()
         .with(predicate::eq(args))
@@ -135,8 +192,8 @@ async fn create_event() {
     let resp = server
         .events_post(
             models::Event {
-                id: event_id_str,
-                data: event_data,
+                id: "".to_string(), // todo(stbrody): remove id arg
+                data: event_data.to_string(),
             },
             &Context,
         )
@@ -184,7 +241,7 @@ async fn register_interest_sort_value() {
         )
         .times(1)
         .returning(|_, _| Ok(true));
-    let mock_model = MockReconModelTest::new();
+    let mock_model = MockAccessModelStoreTest::new();
     let server = Server::new(peer_id, network, mock_interest, Arc::new(mock_model));
     let interest = models::Interest {
         sep: "model".to_string(),
@@ -206,7 +263,7 @@ async fn register_interest_sort_value_bad_request() {
 
     // Setup mock expectations
     let mock_interest = MockReconInterestTest::new();
-    let mock_model = MockReconModelTest::new();
+    let mock_model = MockAccessModelStoreTest::new();
     let server = Server::new(peer_id, network, mock_interest, Arc::new(mock_model));
     let interest = models::Interest {
         sep: "model".to_string(),
@@ -257,7 +314,7 @@ async fn register_interest_sort_value_controller() {
         )
         .times(1)
         .returning(|_, _| Ok(true));
-    let mock_model = MockReconModelTest::new();
+    let mock_model = MockAccessModelStoreTest::new();
     let server = Server::new(peer_id, network, mock_interest, Arc::new(mock_model));
     let resp = server
         .interests_sort_key_sort_value_post(
@@ -312,7 +369,7 @@ async fn register_interest_value_controller_stream() {
         )
         .times(1)
         .returning(|_, _| Ok(true));
-    let mock_model = MockReconModelTest::new();
+    let mock_model = MockAccessModelStoreTest::new();
     let server = Server::new(peer_id, network, mock_interest, Arc::new(mock_model));
     let resp = server
         .interests_sort_key_sort_value_post(
@@ -357,7 +414,7 @@ async fn get_events_for_interest_range() {
             */
     let mock_interest = MockReconInterestTest::new();
     let expected = BuildResponse::event(cid.clone(), vec![]);
-    let mut mock_model = MockReconModelTest::new();
+    let mut mock_model = MockAccessModelStoreTest::new();
     mock_model
         .expect_range_with_values()
         .with(
@@ -411,7 +468,7 @@ async fn test_events_event_id_get_by_event_id_success() {
     let event_id_str = multibase::encode(Base::Base16Lower, event_id.to_bytes());
     let event_data = b"event data".to_vec();
     let event_data_base64 = multibase::encode(multibase::Base::Base64, &event_data);
-    let mut mock_model = MockReconModelTest::new();
+    let mut mock_model = MockAccessModelStoreTest::new();
     mock_model
         .expect_value_for_order_key()
         .with(predicate::eq(event_id))
@@ -439,7 +496,7 @@ async fn test_events_event_id_get_by_cid_success() {
         Cid::from_str("baejbeihyr3kf77etqdccjfoc33dmko2ijyugn6qk6yucfkioasjssz3bbu").unwrap(); // cspell:disable-line
     let event_data = b"event data".to_vec();
     let event_data_base64 = multibase::encode(multibase::Base::Base64, &event_data);
-    let mut mock_model = MockReconModelTest::new();
+    let mut mock_model = MockAccessModelStoreTest::new();
     mock_model
         .expect_value_for_cid()
         .with(predicate::eq(event_cid))
